+++ R/ui_elements.R
@@ -11,6 +11,30 @@ use_telemetry <- function(id = "") {
   if (!is.null(id) && !identical(trimws(id), "")) {
     shiny_namespace <- shiny::NS(trimws(id), "")
   }
+  # Only inject error script for versions of Shiny that don't yet support `onUnhandledError`
+  track_error_script <- if (!("onUnhandledError" %in% ls(getNamespace("shiny")))) {
+    shiny::tags$script(
+      shiny::HTML(
+        glue::glue("
+  $(document).on('shiny:error', function(event) {{
+    var errorData = {{
+      output_id: event.name,
+      message: event.error.message,
+      type: 'error'
+    }};
+    Shiny.setInputValue(
+      '{shiny_namespace}track_error_telemetry_js',
+      errorData,
+      {{priority: 'event'}}
+    );
+  }});
+  "
+        )
+      )
+    )
+  } else {
+    NULL
+  }
   shiny::singleton(shiny::tagList(
     shiny::tags$script(
       type = "text/javascript",
@@ -46,7 +70,8 @@ use_telemetry <- function(id = "") {
       src = "js",
       package = "shiny.telemetry",
       script = "manage-cookies.js"
-    )
+    ),
+    track_error_script
   )
   )
 }