+++ NEWS.md
@@ -1,7 +1,8 @@
-# shiny.telemetry (development version)
+# shiny.telemetry 0.3.0
 
 ### New Features
 
+- Added shiny error tracking (activated by default with `start_session`) (#116).
 - Updated `get_user` method to retrieve user in `shinyproxy` environment (#124).
 - Added flexibility to select between [`RPostgreSQL`, `RPostgres`] drivers (#147).
 - Improved input tracking by implementing inclusion and exclusion logic (#30).
@@ -13,6 +14,7 @@
 - Updates documentation to use markdown format (#153).
 - Improves SQL injection safeguards via `glue::glue_sql` to generated SQL queries (#34).
 - Show proper error message when no telemetry data is available (#177).
+- Adds how-to guides to site (#179 and #180)
 
 ### Bug Fixes